Fixes to nagios plugin perf_data format
It turns out that LibreNMS does not like the original format, but needs it in
the more verbose format. This involves heavy refactoring of the original code
which was more confusing that necessary, having grown organically while learning
the Nagios plugin format.

URL: https://nagios-plugins.org/doc/guidelines.html#AEN200
Signed-off-by: Samuli Seppänen <samuli.seppanen@puppeteers.fi>